backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyPDF2==3.0.1
sentence-transformers==5.1.0
nltk==3.9.1
qdrant-client==1.13.3
openai
anthropic
python-dotenv==1.0.1
tqdm==4.67.1
pytest==8.4.1
